--- src/vuex/modules/list/mutation-types.js
@@ -1,4 +1,0 @@
-export const DELETE_DOCUMENT = 'DELETE_DOCUMENT'
-export const DELETE_DOCUMENTS = 'DELETE_DOCUMENTS'
-export const RECEIVE_DOCUMENTS = 'RECEIVE_DOCUMENTS'
-export const SET_BASIC_FILTER = 'SET_BASIC_FILTER'
