+++ src/vuex/store.js
@@ -3,7 +3,6 @@ import Vuex from 'vuex'
 import auth from './modules/auth/store'
 import common from './modules/common/store'
 import data from './modules/data/store'
-import list from './modules/list/store'
 import crudlDocument from './modules/common/crudlDocument/store'
 
 Vue.use(Vuex)
@@ -12,7 +11,6 @@ export default new Vuex.Store({
   modules: {
     common,
     auth,
-    list,
     data,
     crudlDocument
   },
